last upd

+++ project/spec 2.0/my_malloc.c
@@ -1,7 +1,8 @@
 #pragma once
 
 #include <stdlib.h>
+#include <stdio.h>
 
-void* my_malloc(int n) {
+void* my_malloc(long long n) {
     return malloc(n);
 }
